+++ praktikumsplaner-frontend/frontend/src/views/praktikumsplaetze/MeldungStudium.vue
@@ -268,30 +268,10 @@
                 </v-row>
                 <v-row>
                     <v-col>
-                        <v-select
+                        <MeldezeitraumSelect
                             v-model="praktikumsstelle.meldezeitraumID"
-                            label="Meldezeitraum*"
-                            :items="meldezeitraeume"
-                            :menu-props="customMenuProps"
-                            item-value="id"
-                            item-text="zeitraumName"
-                            outlined
-                        >
-                            <template #item="data">
-                                {{ data.item.zeitraumName }}:
-                                {{
-                                    formatter.formatDateFromString(
-                                        data.item.zeitraum.startZeitpunkt
-                                    )
-                                }}
-                                -
-                                {{
-                                    formatter.formatDateFromString(
-                                        data.item.zeitraum.endZeitpunkt
-                                    )
-                                }}
-                            </template>
-                        </v-select>
+                            :meldezeitraueme="meldezeitraeume"
+                        ></MeldezeitraumSelect>
                     </v-col>
                     <v-col cols="2" />
                     <v-col> </v-col>
@@ -331,10 +311,10 @@ import MeldezeitraumService from "@/api/MeldezeitraumService";
 import PageTitle from "@/components/common/PageTitle.vue";
 import { Programmierkenntnisse } from "@/types/YesNoEgalBoolean";
 import "@/directives/security";
-import { useFormatter } from "@/composables/formatter";
 import { useUserStore } from "@/stores/user";
 import { APP_SECURITY } from "@/constants";
 import Meldezeitraum from "@/types/Meldezeitraum";
+import MeldezeitraumSelect from "@/components/praktikumsplaetze/Meldung/MeldezeitraumSelect.vue";
 
 const activeMeldezeitraum = ref<boolean>(false);
 
@@ -346,7 +326,9 @@ const zuweisungsZeitraum = ref<string>("");
 const isAusbildungsleitung = ref<boolean>(false);
 const userStore = useUserStore();
 const validationRules = useRules();
-const formatter = useFormatter();
+const customMenuProps = {
+    offsetY: true,
+};
 const requiredRule = [validationRules.notEmptyRule("Darf nicht leer sein.")];
 const emailRule = [
     validationRules.notEmptyRule("Darf nicht leer sein."),
@@ -387,9 +369,6 @@ const taetigkeitenRule = [
     ),
 ];
 
-const customMenuProps = {
-    offsetY: true,
-};
 const form = ref<HTMLFormElement>();
 const meldezeitraeume = computed(() => {
     return [
